# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  empresa.criaEmpresaRequest:
    properties:
      cnpj:
        example: 12.345.678/0001-95
        type: string
      nomeFantasia:
        example: Minha Empresa
        type: string
      razaoSocial:
        example: Minha Empresa LTDA
        type: string
    required:
    - cnpj
    - nomeFantasia
    - razaoSocial
    type: object
...
  localidade.createRequest:
    properties:
      cep:
        type: string
      empresa_id:
        type: integer
      nome:
        type: string
      raio_geofence_metros:
        type: number
    required:
    - cep
    - empresa_id
    - nome
    - raio_geofence_metros
    type: object
...
  model.Empresa:
    properties:
      cnpj:
        type: string
      id:
        type: integer
      localidades:
        items:
          $ref: '#/definitions/model.Localidade'
        type: array
      nome_fantasia:
        type: string
      razao_social:
        type: string
    type: object
...
  model.Localidade:
    properties:
      bairro:
        type: string
      cep:
        type: string
      cidade:
        type: string
      empresa_id:
        type: integer
      estado:
        type: string
      id:
        type: integer
      latitude:
        type: number
      logradouro:
        type: string
      longitude:
        type: number
      nome:
        description: 'Ex: "Matriz São Paulo", "Filial Rio"'
        type: string
      numero:
        type: string
      raio_geofence_metros:
        type: number
    type: object
...
host: localhost:8083
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API de alta performance para um sistema de Ponto Eletrônico.
  license:
    name: MIT
    url: https://github.com/Loviiin/ponto-api-go/blob/main/LICENSE
  termsOfService: http://swagger.io/terms/
  title: Ponto API em Go
  version: "1.0"
paths:
...
  /empresas:
    get:
      description: Retorna uma lista de todas as empresas cadastradas. Requer autenticação.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Empresa'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Lista todas as empresas
      tags:
      - Empresas
    post:
      consumes:
      - application/json
      description: Registra uma nova empresa no sistema e configura cargos e permissões
        padrão para ela.
      parameters:
      - description: Dados da nova empresa
        in: body
        name: empresa
        required: true
        schema:
          $ref: '#/definitions/empresa.criaEmpresaRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Empresa'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Cria uma nova empresa
      tags:
      - Empresas
  /empresas/{id}:
    delete:
      description: Deleta uma empresa permanentemente. Requer permissão 'DELETAR_EMPRESA'.
      parameters:
      - description: ID da Empresa
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Deleta uma empresa
      tags:
      - Empresas
    get:
      description: Retorna os dados de uma empresa específica pelo seu ID.
      parameters:
      - description: ID da Empresa
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Empresa'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Busca uma empresa por ID
      tags:
      - Empresas
    put:
      consumes:
      - application/json
      description: Atualiza os dados de uma empresa. Requer permissão 'EDITAR_EMPRESA'.
      parameters:
      - description: ID da Empresa
        in: path
        name: id
        required: true
        type: integer
      - description: Dados para atualização
        in: body
        name: dados
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Atualiza uma empresa
      tags:
      - Empresas
  /empresas/{id}/localidades:
    get:
      description: Retorna uma lista de todas as localidades de uma empresa. Requer
        permissão 'GERENCIAR_LOCALIDADES'.
      parameters:
      - description: ID da Empresa
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Localidade'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Lista as localidades de uma empresa
      tags:
      - Localidades
...
  /localidades:
    post:
      consumes:
      - application/json
      description: Cria uma nova localidade (matriz ou filial) para uma empresa, buscando
        o endereço e as coordenadas a partir do CEP. Requer permissão 'GERENCIAR_LOCALIDADES'.
      parameters:
      - description: Dados da nova localidade
        in: body
        name: localidade
        required: true
        schema:
          $ref: '#/definitions/localidade.createRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Localidade'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Cria uma nova localidade
      tags:
      - Localidades
...
securityDefinitions:
  BearerAuth:
    description: '"Digite ''Bearer'' seguido de um espaço e o seu token."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"